+++ orionadapter/sec/authz/xacml/response.go
@@ -1,4 +1,4 @@
-package authz
+package xacml
 
 import (
 	"encoding/xml"
