+++ orionadapter/sec/authz/xacml/response_test.go
@@ -1,4 +1,4 @@
-package authz
+package xacml
 
 import (
 	"testing"
